fix(callapi-legacy,callapi): update resolveHeaders function and add requestKey option - Enhance resolveHeaders function to handle various types of auth and body options - Add requestKey option to callApi function in dev/src/client.ts
