+++ dev/src/client.ts
@@ -9,6 +9,7 @@ const [foo1, foo2, foo3, foo4] = await Promise.all([
 	callApi("https://dummyjson.com/products/:id", {
 		method: "GET",
 		params: [1],
+		requestKey: "hello",
 	}),
 
 	callApi("https://dummyjson.com/products/:id", {
