+++ packages/callapi-legacy/package.json
@@ -1,7 +1,7 @@
 {
 	"name": "@zayne-labs/callapi-legacy",
 	"type": "module",
-	"version": "1.0.0-rc.2",
+	"version": "1.0.0-rc.4",
 	"description": "A lightweight wrapper over fetch with quality of life improvements like built-in request cancellation, retries, interceptors and more",
 	"author": "Ryan Zayne",
 	"license": "MIT",
